third_party/blink/web_tests/virtual/scoped-custom-element-registry-disabled/external/wpt/custom-elements/scoped-registry/ShadowRoot-innerHTML-upgrade.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Upgrade into autonomous custom element when inserted via innerHTML
  Illegal constructor
[FAIL] Upgrade into autonomous custom element when definition is added
  Illegal constructor
[FAIL] Upgrade into customized built-in element when inserted via innerHTML
  Illegal constructor
[FAIL] Upgrade into customized built-in element when definition is added
  Illegal constructor
[FAIL] Upgrade into autonomous custom element should not inherit from global registry for missing values
  Illegal constructor
Harness: the test ran to completion.
